Simplify the code
In my first version ,when the fibonacci function get the argument equal to 0 or 1 it will jump to .L3 or .L4 to return 0 or 1.In this version ,I let r5 contain value -1 and let r6 contain value 1 .First time r5 add r6 ,we get 0.Second time r5 add r6 we get 1. This is the same as the first and second number of Fibonacii sequence so that I don't have to determine if the argument is 0 or 1 at first.